add article delete update

+++ blog/app/Http/Controllers/ArticleController.php
@@ -3,10 +3,17 @@
 namespace App\Http\Controllers;
 
 use App\Models\Article;
+use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'detail']);
+    }
+
     public function index()
     {
         $data = Article::latest()->paginate(5);
@@ -27,10 +34,7 @@ class ArticleController extends Controller
 
     public function add()
     {
-        $data = [
-            ["id" => 1, "name" => "News"],
-            ["id" => 2, "name" => "Techs"],
-        ];
+        $data = Category::all();
 
         return view('articles.add',[
             'categories' => $data
@@ -53,16 +57,53 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = auth()->user()->id;
         $article->save();
 
         return redirect('/articles');
     }
 
+    public function edit($id)
+    {
+        $categories = Category::all();
+        $article = Article::find($id);
+
+        return view('articles.edit',[
+           'categories' => $categories,
+            'article' => $article
+        ]);
+    }
+
+    public function update($id)
+    {
+        $validator = validator(request()->all(),[
+            'title' => 'required',
+            'body' => 'required',
+            'category_id' => 'required'
+        ]);
+
+        if ($validator->fails()) {
+            return back()->withErrors($validator);
+        }
+
+        $article = Article::find($id);
+        $article->title = request()->title;
+        $article->body = request()->body;
+        $article->category_id = request()->category_id;
+        $article->save();
+
+        return redirect("/articles/detail/$id");
+    }
+
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
 
-        return redirect('/articles')->with('info', 'Article deleted');
+        if(Gate::allows('delete-article', $article)) {
+            $article->delete();
+            return redirect('/articles')->with('info', 'Deleted an article');
+        }
+
+        return back()->with('info', 'Unauthorized');
     }
 }
