+++ blog/app/Http/Controllers/CommentController.php
@@ -2,16 +2,32 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Comment;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
+use App\Models\Comment;
 
 class CommentController extends Controller
 {
+    public function __construct()
+    {
+        $this->middleware('auth');
+    }
+
     public function create()
     {
-        $comment = new Comment();
+        $validator = validator(request()->all(), [
+            "content" => "required",
+            "article_id" => "required",
+        ]);
+
+        if($validator->fails()) {
+            return back()->withErrors($validator);
+        }
+
+        $comment = new Comment;
         $comment->content = request()->content;
         $comment->article_id = request()->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return back();
@@ -20,8 +36,12 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back();
+        if(Gate::allows('delete-comment', $comment)) {
+            $comment->delete();
+            return back();
+        }
+
+        return back()->with('info', 'Unauthorize');
     }
 }
